Progress on validation, mods to hashmap
The "name" attribute of the CSV is now the location's standard name. Previously, it was normalized. We want the unnormalized name to be added to the list.

Right now, if we find a matching location in the hashmap, we add it to the accepted locations list. If the player enters a wrong guess, the location is no longer added to the list. But, the current player still is allowed to play a next guess beginning with the nex letter.