Now, valid guesses add the matching markers on the map.
Just thought of an interesting data problem. We might run into cases where we have a location name that refers to more than one real location. E.g. Naples, Florida vs. Naples Italy. Currently, the way data is processed, only one row in the location CSV is Naples. Extend the value to be a list of  CCS's matching the name, and when it is guessed, add all of them to the map?